+++ content/template/listTemplate/template_A_listCourses.php
@@ -16,24 +16,28 @@ if($categoryID > 0)
 {
 	$f = new XFilter('CategoryID', '=', $categoryID);
 	$filtering->AddItem($f);
+	$attributes['category'] = $categoryID;
 }
 
 if(isset($_REQUEST['eduadmin-category']))
 {
 	$f = new XFilter('CategoryID', '=', $_REQUEST['eduadmin-category']);
 	$filtering->AddItem($f);
+	$attributes['category'] = $_REQUEST['eduadmin-category'];
 }
 
 if(isset($_REQUEST['eduadmin-city']))
 {
 	$f = new XFilter('LocationID', '=', $_REQUEST['eduadmin-city']);
 	$filtering->AddItem($f);
+	$attributes['city'] = $_REQUEST['eduadmin-city'];
 }
 
 if(isset($_REQUEST['eduadmin-subject']))
 {
 	$f = new XFilter('SubjectID', '=', $_REQUEST['eduadmin-subject']);
 	$filtering->AddItem($f);
+	$attributes['subject'] = $_REQUEST['eduadmin-subject'];
 }
 
 if(!empty($filterCourses))
@@ -106,6 +110,7 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 
 if(isset($_REQUEST['eduadmin-level']) && !empty($_REQUEST['eduadmin-level']))
 {
+	$attributes['courselevel'] = $_REQUEST['eduadmin-level'];
 	$edo = array_filter($edo, function($object) {
 		$cl = get_transient('eduadmin-courseLevels');
 		foreach($cl as $subj)
@@ -146,6 +151,7 @@ if(isset($_REQUEST['eduadmin-category']))
 {
 	$f = new XFilter('CategoryID', '=', $_REQUEST['eduadmin-category']);
 	$filtering->AddItem($f);
+	$attributes['category'] = $_REQUEST['eduadmin-category'];
 }
 
 if(isset($_REQUEST['eduadmin-subject']))
@@ -215,6 +221,8 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 		set_transient('eduadmin-subjects', $subjects, DAY_IN_SECONDS);
 	}
 
+	$attributes['subject'] = $_REQUEST['eduadmin-subject'];
+
 	$ede = array_filter($ede, function($object) {
 		$subjects = get_transient('eduadmin-subjects');
 		foreach($subjects as $subj)
@@ -230,6 +238,7 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 
 if(isset($_REQUEST['eduadmin-level']) && !empty($_REQUEST['eduadmin-level']))
 {
+	$attributes['courselevel'] = $_REQUEST['eduadmin-level'];
 	$ede = array_filter($ede, function($object) {
 		$cl = get_transient('eduadmin-courseLevels');
 		foreach($cl as $subj)
@@ -322,9 +331,11 @@ if(!empty($edo))
 
 	$showEventVenue = get_option('eduadmin-showEventVenueName', false);
 
-	?><div style="display: none;" class="eduadmin-courselistoptions" data-subject="<?php echo @esc_attr($attributes['subject']); ?>"
+	?><div style="display: none;" class="eduadmin-courselistoptions"
+	data-subject="<?php echo @esc_attr($attributes['subject']); ?>"
 	data-category="<?php echo @esc_attr($attributes['category']); ?>"
 	data-city="<?php echo @esc_attr($attributes['city']); ?>"
+	data-courselevel="<?php echo @esc_attr($attributes['courselevel']); ?>"
 	data-spotsleft="<?php echo @get_option('eduadmin-spotsLeft', 'exactNumbers'); ?>"
 	data-spotsettings="<?php echo @get_option('eduadmin-spotsSettings', "1-5\n5-10\n10+"); ?>"
 	data-fewspots="<?php echo @get_option('eduadmin-alwaysFewSpots', "3"); ?>"