+++ content/template/listTemplate/template_A_listEvents.php
@@ -14,6 +14,7 @@ $baseUrl = $surl . '/' . $cat;
 	{
 		$f = new XFilter('CategoryID', '=', $categoryID);
 		$filtering->AddItem($f);
+		$attributes['category'] = $categoryID;
 	}
 
 	if(!empty($filterCourses))
@@ -32,9 +33,10 @@ $baseUrl = $surl . '/' . $cat;
 	{
 		$f = new XFilter('CategoryID', '=', $_REQUEST['eduadmin-category']);
 		$filtering->AddItem($f);
+		$attributes['category'] = $_REQUEST['eduadmin-category'];
 	}
 
-	$edo = $eduapi->GetEducationObject($edutoken, '', $filtering->ToString());
+	$edo = $eduapi->GetEducationObjectV2($edutoken, '', $filtering->ToString(), false);
 }
 
 if(isset($_REQUEST['searchCourses']) && !empty($_REQUEST['searchCourses']))
@@ -112,6 +114,7 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 	{
 		$f = new XFilter('LocationID', '=', $_REQUEST['eduadmin-city']);
 		$filtering->AddItem($f);
+		$attributes['city'] = $_REQUEST['eduadmin-city'];
 	}
 
 	if(isset($_REQUEST['eduadmin-subject']))
@@ -124,6 +127,8 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 	{
 		$f = new XFilter('CategoryID', '=', $_REQUEST['eduadmin-category']);
 		$filtering->AddItem($f);
+
+		$attributes['category'] = $_REQUEST['eduadmin-category'];
 	}
 
 	$f = new XFilter('CustomerID','=','0');
@@ -170,6 +175,8 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 		set_transient('eduadmin-subjects', $subjects, DAY_IN_SECONDS);
 	}
 
+	$attributes['subject'] = $_REQUEST['eduadmin-subject'];
+
 	$ede = array_filter($ede, function($object) {
 		$subjects = get_transient('eduadmin-subjects');
 		foreach($subjects as $subj)
@@ -185,6 +192,7 @@ if(isset($_REQUEST['eduadmin-subject']) && !empty($_REQUEST['eduadmin-subject'])
 
 if(isset($_REQUEST['eduadmin-level']) && !empty($_REQUEST['eduadmin-level']))
 {
+	$attributes['courselevel'] = $_REQUEST['eduadmin-level'];
 	$ede = array_filter($ede, function($object) {
 		$cl = get_transient('eduadmin-courseLevels');
 		foreach($cl as $subj)
@@ -310,6 +318,7 @@ if(isset($_REQUEST['searchCourses']) && !empty($_REQUEST['searchCourses']))
 	data-template="A"
 	data-subject="<?php echo @esc_attr($attributes['subject']); ?>"
 	data-category="<?php echo @esc_attr($attributes['category']); ?>"
+	data-courselevel="<?php echo @esc_attr($attributes['courselevel']); ?>"
 	data-city="<?php echo @esc_attr($attributes['city']); ?>"
 	data-spotsleft="<?php echo @get_option('eduadmin-spotsLeft', 'exactNumbers'); ?>"
 	data-spotsettings="<?php echo @get_option('eduadmin-spotsSettings', "1-5\n5-10\n10+"); ?>"
@@ -374,7 +383,7 @@ foreach($ede as $object)
 			"</div>";
 		}
 
-		if($showEventPrice) {
+		if($showEventPrice && isset($object->Price)) {
 			echo "<div class=\"priceInfo\">" . sprintf(edu__('From %1$s'), convertToMoney($object->Price, $currency)) . " " . edu__($incVat ? "inc vat" : "ex vat") . "</div> ";
 		}